refactor: merge useMenubar and useTray into useNativeUI hook
Consolidate two hooks that subscribed to the same 5 timer store selectors into one. Single useEffect handles both menubar title/tooltip and tray tooltip updates.
